+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 
-- Check if `sub` claim matches the status list-uri in the cache to fail fast in addition to the check in the verification process (#1207)
+- Check if `sub` claim matches the status list-uri in the cache to fail fast in addition to the check in the verification process `(#1207)`
+- Check if trust statement issuer matches the issuer of the status list `(#1210)`
 
 # [4.2.0] - 2026-08-21